Revert ":alien: Migrate rand from 0.8.5 to 0.9.0 and rand_chacha from 0.3.1 to 0.9.0"
This reverts commit 8ef399c15f67ca8f3a40ce51b5acd015a02b08d0.

+++ cli/src/command/update.rs
@@ -260,7 +260,7 @@ fn update_archive<Strategy: TransformStrategy>(args: UpdateCommand) -> io::Resul
 
     let (tx, rx) = std::sync::mpsc::channel();
 
-    let random = rand::random::<u64>();
+    let random = rand::random::<usize>();
     let temp_dir_path = temp_dir().unwrap_or_else(|| {
         archive_path
             .parent()
